+++ src/lib.rs
@@ -1,7 +1,7 @@
 pub mod model;
 pub mod parser;
 
-pub use model::{JsonParseError, JsonValue};
+pub use model::{JsonParseError, JsonParseOptions, JsonValue};
 pub use parser::{
     parse_array, parse_bool, parse_json, parse_null, parse_number, parse_object, parse_string,
 };